pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-cube"
version = "1.1.0"
description = "Autonomous multi-agent coding workflow orchestration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Aetheron", email = "dev@aetheron.ai"}
]
keywords = ["ai", "agents", "coding", "automation", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "aiofiles>=23.0.0",
    "gitpython>=3.1.0",
    "typing-extensions>=4.0.0",
    "pyyaml>=6.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
cube = "cube.cli:app"

[project.urls]
Homepage = "https://github.com/aetheronhq/agent-cube"
Repository = "https://github.com/aetheronhq/agent-cube"
Issues = "https://github.com/aetheronhq/agent-cube/issues"

[tool.hatch.build.targets.wheel]
packages = ["python/cube"]

[tool.hatch.build.targets.sdist]
include = [
    "python/cube/**/*.py",
    "README.md",
    "LICENSE",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_optional = true
